+++ back/luckybocky/src/main/java/com/project/luckybocky/user/entity/User.java
@@ -5,9 +5,12 @@ import com.project.luckybocky.common.BaseEntity;
 import com.project.luckybocky.feedback.entity.Feedback;
 import com.project.luckybocky.pocket.entity.Pocket;
 import com.project.luckybocky.report.entity.Report;
+import com.project.luckybocky.user.dto.UserInfoDto;
+
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,61 +21,76 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class User extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    //12-23 창희 JoinColumn시 칼럼매핑을 하지못해서 명시적으로 추가
-    @Column(name ="user_seq")
-    private Integer userSeq;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	//12-23 창희 JoinColumn시 칼럼매핑을 하지못해서 명시적으로 추가
+	@Column(name = "user_seq")
+	private Integer userSeq;
 
-    @Column(unique = true)
-    private String userKey;
+	@Column(unique = true)
+	private String userKey;
 
-    @Column(nullable = true)
-    private String userNickname;
+	@Column(nullable = true)
+	private String userNickname;
 
-//    @Column(unique = true)
-    private String firebaseKey;
+	//    @Column(unique = true)
+	private String firebaseKey;
 
-    @Column(nullable = false)
-    private boolean alarmStatus;
+	@Column(nullable = false)
+	private boolean alarmStatus;
 
-    @Column(nullable = false)
-    private boolean fortuneVisibility;
+	@Column(nullable = false)
+	private boolean fortuneVisibility;
 
-    @OneToMany(mappedBy = "user")
-    List<Article> articles = new ArrayList<>();
+	@OneToMany(mappedBy = "user")
+	List<Article> articles = new ArrayList<>();
 
-    @OneToMany(mappedBy = "user")   // cascade 설정 X
-    List<Pocket> pockets = new ArrayList<>();
+	@OneToMany(mappedBy = "user")   // cascade 설정 X
+	List<Pocket> pockets = new ArrayList<>();
 
+	//12-23 창희 피드백, 신고 칼럼 추가 start
+	@OneToMany(mappedBy = "user")
+	private List<Feedback> feedbacks = new ArrayList<>();
 
-    //12-23 창희 피드백, 신고 칼럼 추가 start
-    @OneToMany(mappedBy = "user")
-    private List<Feedback> feedbacks = new ArrayList<>();
+	@OneToMany(mappedBy = "reporter")
+	private List<Report> reporters; // 내가 신고한 목록
 
-    @OneToMany(mappedBy = "reporter")
-    private List<Report> reporters; // 내가 신고한 목록
+	@OneToMany(mappedBy = "offender")
+	private List<Report> offenders; // 내가 신고당한 목록
+	//12-23 창희 피드백, 신고 칼럼 추가 end
 
-    @OneToMany(mappedBy = "offender")
-    private List<Report> offenders; // 내가 신고당한 목록
-    //12-23 창희 피드백, 신고 칼럼 추가 end
+	@Override
+	public String toString() {
+		return "User{" +
+			"userSeq=" + userSeq +
+			", userKey='" + userKey + '\'' +
+			", userNickname='" + userNickname + '\'' +
+			", firebaseKey='" + firebaseKey + '\'' +
+			", alarmStatus=" + alarmStatus +
+			", fortuneVisibility=" + fortuneVisibility +
+			'}';
+	}
 
-    @Override
-    public String toString() {
-        return "User{" +
-                "userSeq=" + userSeq +
-                ", userKey='" + userKey + '\'' +
-                ", userNickname='" + userNickname + '\'' +
-                ", firebaseKey='" + firebaseKey + '\'' +
-                ", alarmStatus=" + alarmStatus +
-                ", fortuneVisibility=" + fortuneVisibility +
-                '}';
-    }
+	public void updateUserInfo(String userNickname, boolean alarmStatus, boolean fortuneVisibility) {
+		setUserNickname(userNickname);
+		setAlarmStatus(alarmStatus);
+		setFortuneVisibility(fortuneVisibility);
+	}
 
-    public void updateUserInfo(String userNickname, boolean alarmStatus, boolean fortuneVisibility){
-        setUserNickname(userNickname);
-        setAlarmStatus(alarmStatus);
-        setFortuneVisibility(fortuneVisibility);
-    }
+	//=====창희 dto 함수 start
+
+	public UserInfoDto getUserInfo(User user) {
+
+		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+		String formattedDate = user.getCreatedAt().format(formatter);
+
+		return UserInfoDto.builder()
+			.userNickname(user.getUserNickname())
+			.alarmStatus(user.isAlarmStatus())
+			.fortuneVisibility(user.isFortuneVisibility())
+			.createdAt(formattedDate)
+			.build();
+	}
+	//=====창희 dto 함수 end
 
 }
