+++ back/luckybocky/src/main/java/com/project/luckybocky/user/exception/UserNotFoundException.java
@@ -4,7 +4,9 @@ import org.springframework.http.HttpStatus;
 
 import com.project.luckybocky.common.CustomException;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 public class UserNotFoundException extends CustomException {
@@ -13,4 +15,5 @@ public class UserNotFoundException extends CustomException {
     public UserNotFoundException(){
         super(statusCode,message);
     }
+
 }
